isolate wrapped_function to a dedicated Wrapper class.
