+++ src/taipy/core/_orchestrator/_dispatcher/_development_job_dispatcher.py
@@ -44,5 +44,5 @@ class _DevelopmentJobDispatcher(_JobDispatcher):
         """
         config_as_string = _TomlSerializer()._serialize(Config._applied_config)
 
-        rs = self._run_wrapped_function(Config.job_config.mode, config_as_string, job.id, job.task)
+        rs = self._wrapped_function(Config.job_config.mode, config_as_string, job.id, job.task)
         self._update_job_status(job, rs)
