+++ src/taipy/core/_orchestrator/_dispatcher/_standalone_job_dispatcher.py
@@ -39,7 +39,7 @@ class _StandaloneJobDispatcher(_JobDispatcher):
 
         config_as_string = _TomlSerializer()._serialize(Config._applied_config)
         future = self._executor.submit(
-            self._run_wrapped_function, Config.job_config.mode, config_as_string, job.id, job.task
+            self._wrapped_function, Config.job_config.mode, config_as_string, job.id, job.task
         )
 
         self._set_dispatched_processes(job.id, future)  # type: ignore
